virttest.remote: Change default log_function for remote_login
The old remote_login used a `utils_logfile.log_line` log_function by
default, making it convenient to set the output log file. Let's bring
back the default by changing `None` to our log function.

This can change the log_function in case people explicitly select
log_function to None but it should not matter as the function is only
used when `log_filename` is set. So the only effect would be when one
chooses to set the `log_filename` but deliberately wanted to disable
logging via the log_function, which should be very rare.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>