Support image convert disable
This is to make the command 'qemu-img convert' optional for netfs.
Because image_convert is True as default, this change will not
impact the existing invoke codes. The reason to change this is
because in most cases there is no need to do the image convert
which increases the migration duration. After this, the user
invokes set_vm_disk() with image_convert='no' to avoid converting
the image.

Signed-off-by: Dan Zheng <dzheng@redhat.com>