Remove installation and package availability based on Python 2
It's been decided that Python 2 will be dropped.  That doesn't necessarily
mean we'll go through all the code and remove everything that is valid on
Python 2, but that we won't ship or support Avocado-VT running on Python 2.

This removes the conditionals for requirements on Python 2, and removes the
(already optional) building of Python 2 RPMs.

Reference: https://github.com/avocado-framework/avocado-vt/issues/2963
Signed-off-by: Cleber Rosa <crosa@redhat.com>