Customize tls setup
Before this PR, we always scp all required files like cakey, cacert,
severkey ... to the TLS server host and client host which are normal
way for TLS connection setup. But sometimes we need partial setup.
So this PR is to enable the capability to allow the user to define
which files will be scped to the TLS server/client host.

Signed-off-by: Dan Zheng <dzheng@redhat.com>