Extra Params: restore support for command line --vt-extra-params
The action being given to the "--vt-extra-params", by omission, was
"store".  That caused every entry to override the previous one.  What
we really need, to have a list of options, is "append".

While at it, let's change the number of arguments, because it does
not make any sense to have "--vt-extra-params" with zero arguments.

Signed-off-by: Cleber Rosa <crosa@redhat.com>