Default serial run for vt nrunner
The avocado plan for near future to change default runner form legacy to
nrunner. It is necessary to prepare avocado-vt for this change. The vt
nrunner doesn't support parallel run, and it is necessary to set
`--nrunner-max-parallel-tasks` to 1 before each run with nrunner. This
might be frustrating when the nrunner becomes the default  avocado
runner. This commit sets  `--nrunner-max-parallel-tasks` to 1 as default
when the avocado-vt is bootstrapped. So it won’t be necessary to set that
variable each time the avocado-vt is run.

Reference: #3030
Signed-off-by: Jan Richter <jarichte@redhat.com>