Redefine the logger to avoid excessive output at the LOG_UI
As we discussed at avocado-framework/avocado#4797 and #3169, the current
logger will display a lot of excessive output on the UI. These outputs
are the same as debug.log, they should only be in logs if we don't use
the verbose mode, so this is unexpected behavior. Beraldo Leal provides
a good solution in #3169, the code works for me, so I borrowed his
solution to solve this problem accordingly.

Signed-off-by: Beraldo Leal <bleal@redhat.com>
Signed-off-by: Yihuang Yu <yihyu@redhat.com>