libvirt: support search stdout on error
Usually we check the expected error messages in stderr. But some
virsh command also outputs the useful message to stdout when the
exit status is not zero. For example,
virsh command: hypervisor-cpu-compare /tmp/log
status: 1
stdout: CPU described in /tmp/log is incompatible with the CPU
        provided by hypervisor on the host
stderr:

So this is to support to check stdout, too.

Signed-off-by: Dan Zheng <dzheng@redhat.com>