avocado_vt/test.py: remove non vt parameters
First, avocado_params() is a non-standard method is a leftover from
around 6 years ago, and it was last present in Avocado 0.32 (being
removed on c344926c1 part of 0.33).

Second, the actual parameters used to instantiate the VirtTest class
are (from avocado_vt/discovery.py:DiscoveryMixIn.convert_parameters()):

    test_parameters = {'name': test_name,
                       'vt_params': params}

Meaning that "params" is not being given to VirtTest.

Signed-off-by: Cleber Rosa <crosa@redhat.com>