VT-runner with test.VirtTest class
This is the restructuring of avocado vt-runner. This change brings the
reusage of vt legacy code in vt nrunner. Thanks to that, there won't be
any compatibility issues which were reported in #3221, and  also it will
be easier to save all the vt log files into the test-result directory
which fix issue #3213.

Reference: #3213, #3221
Signed-off-by: Jan Richter <jarichte@redhat.com>