libvirt_vm: using module name as logger namespace
As discussed in multiple threads in Avocado project, we are using this
new logging approach to avoid touching on the root logger since July.
The next LTS is counting on that and we need to adapt the logging calls
here to use its own namespace.

This is just an test/experiment to see how things will behave here. I'm
choosing this file because is the one with more logging calls.

Signed-off-by: Beraldo Leal <bleal@redhat.com>