qcontainer: Rename file of vTPM daemon to be unique
There is an issue that multiple vTPM daemoms will use the same tpm
state path and sock patch on both source and destination guest
in the scenario of localhost migration. This will result in datat
conflict. So work around to distinguish though adding the randon
string suffix to the filename to make it unique.

Signed-off-by: Yongxue Hong <yhong@redhat.com>