Bump the aexpect version to 1.6.2
In commit c6d915d3bf08d4cc5693018db4a29f1dc6656020, the aexpect was
bumped to 1.6.3, but aexpect 1.6.3 has problem which blocks virt-v2v
testing on windows guests.

The changes in the defective commit is only applied to some rare cases.
Compared the impact to virt-v2v, it's worth reverting it and wait following
PR to be merged, then bump it to the new aexpect version.

The aexpect PR:
https://github.com/avocado-framework/aexpect/pull/90

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>