Cirrus CI: do not require compatibility with latest Avocado
Avocado has taken some steps to allow faster development of new
features, while still caring for compatibility.

Out of those, the most relevant one is the existance of LTS releases,
which should can receive fixes and be used for a long time with a
predictable experience.

With the release of 92.X, and the the removal of the legacy "runner"
architecture in Avocado (among other changes), it's necessary to
detach Avocado-VT from the very latest Avocado commits.

By "allowing failures" for one specific job, we can monitor the
"compatibility delta" between Avocado and Avocado-VT, and work on
compatibility issues with more time and better planning.

The goal continues being having all Avocado and Avocado-VT regular
releases compatible, but that will depend on available resources.
One compromise scenario may be having at least every "even" release
compatible.  Worst case scenario, is having every new (and current LTS
version compatible).

This is also an invitation for developers to work on this specific are
and make the goal a reality, and if we fail to have compatibility
achieved for every release, we need to understand the resource
limitations and settle on one of the other scenarios (every couple of
releases or LTS releases).

Signed-off-by: Cleber Rosa <crosa@redhat.com>