Workaround for removal of "allow_output_check" parameter
The allow_output_check parameter on the avocado.utils.process.run()
method was meant for use by the "--allow-output-check" feature, but
given it was public API, using it was fair game.

To the best of my knowledge, the parameter is not really needed in
Avocado-VT (it was a nice to have).

Reference: https://github.com/avocado-framework/avocado-vt/issues/3299
Signed-off-by: Cleber Rosa <crosa@redhat.com>