virsh: Remove unnecessary change
We did some changes in 4873a25 to fix error "Not found event tray-change after
40 seconds".

Since the error is hard to be reproduced locally and the error breaks a
lot cases in CI. So we merged it firstly.
After further debug, we noticed one of the changes is wrong, so remove
it.

Fix: 4873a25f Debug draft
Signed-off-by: Liu Yiding <liuyd.fnst@fujitsu.com>