virsh: Remove unnecessary change
We did some changes in 4873a25 to fix error "Not found event tray-change after
40 seconds".
Since the error is hard to be reproduced locally and the error breaks a lot
cases in CI. So we merged that draft PR.

After further debug, we noticed that some changes are unnecessary or
inappropriate. Replacing time.sleep with utils_misc.wait_for is much better.

Fix: 4873a25f Debug draft
Signed-off-by: Liu Yiding <liuyd.fnst@fujitsu.com>