cartesian_config: can't replace variable with names with "-"
Value substitutions of variables with a "-" weren't
being parsed properly due to the eval method used to do so.
Support for arithmetic operations between variables of
the cartesian dict is not documented. So the usage of
eval doesn't seem to be a must. Moreover, its usage can
lead to considerable security risks.
This patch changes the eval use by a directly looking up
into the dictionary.
Unittests have been updated to match up the supported
substitution operations.
This commit reverts the functionality added by
commit 9c24f76f753e706c4c4ee92a791adc75b015910a that
allowed using basic python ops in the configuration
code.

Signed-off-by: bgartzi <bgartzia@redhat.com>