docs/contributing: fix broken links
The current links to the Autotest wiki are broken.  The fixed links
are not ideal, as they refer to the Autotest information, which was
once in sync with Avocado-VT development, but it's not any longer.

Still, it's better to have a functioning link with some of the
information still helpful and relevant, than broken ones.

Signed-off-by: Cleber Rosa <crosa@redhat.com>