utils_net: added additional checks to arp parsing
A try/except clause has been added to check that the data that is read
from the arp file is completely valid, and that there is not any data
missing.
In case the arp entry is invalid for any reason, an OSError is assumed
and so raised giving some related information to the user.

Signed-off-by: bgartzi <bgartzia@redhat.com>