usb: Change usbdev port bus and type parsing keys
The parameters from usb bus and port were not being properly read since
the configured variable name was "usb_bus" or "usb_port" and the
parameter being read was simply "bus" or "port". This commit changes the
variable names to a more clear "usbdev_bus" or "_port" and also changes
the base cfg file. Additionally, this changes the "usb_type" key to
"usbdev_type" for consistency with the other keys.
We are taking this change progressively. In other words, we are only
encouraging the use of "usbdev_*" parameters and throwing a warning if
others are used.

Signed-off-by: bgartzi <bgartzia@redhat.com>