nodedev_xml: Add XML representation for mediated devices
Add mediated device XML representation as node device.

The definition of the `NodedevXMLBase.set_cap` method shows that
in order to construct a valid node device xml we need to
a) construct a `CAPXML`
b) provide an implementation of the `XMLTreeFile.getroot` method
   to substitute the `NodedevXMLBase` instance' xml tree's `<capability>`
   element by the full xml tree as defined by a)

Implement `MdevXML` subclass of `CAPXML` in order to meet a)
Add `getroot` method to `CAPXML` to delegate to the instance's method.

Finally, fix typo in the `NodedevXMLBase.del_cap` method.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>