utils_netperf: Add a workaround to handle Windows path correctly
Currently, `Netperf*` classes use `os.path` to get the "remote"
path of theirs relevant files, however, this is not correct since
the code would be running on local instead of the remote system.

Usually, we need to figure out a proper fix for the issue, but it
looks like a refactoring of the entire `remote.Remote_Package` stuff
is something necessary, and therefore let's firstly have a workaround
for the issue.

Signed-off-by: Xu Han <xuhan@redhat.com>