CI: pin the versions of tp-qemu and tp-libvirt
Up until now, the bootstrap process run on CI will fetch the HEAD of
the repository URIs given.  In testing, it's a good practice to only
change one component at a time.

By changing only one component at time, we are able to quickly point
at compatibility regressions (say a tp-qemu test needs something that
is being dropped or changing behavior in Avocado-VT), or, be required
to update the reference on the test provider to a version that can
make use of a new or improved Avocado-VT API.

In order to expand the integration coverage of the test providers and
Avocado-VT, it's my understanding that this locking of versions is
needed so that maintainers can keep their sanity.

Signed-off-by: Cleber Rosa <crosa@redhat.com>