CI: remove the waiver of failures with latest Avocado
This marks the point where Avocado-VT and Avocado are compatible (to
the extent that is being tested here).  With this, an Avocado-VT
release matching an Avocado release will be produced.

In short, from the latest Avocado (tested with here) there will be a
release 98.0, and from the latest Avocado-VT commit there will be
produced a release 98.0 too.  Then both Avocado and Avocado-VT 98.0
will be compatible releases between themselves.

There are two ways to deal with the waiver after the releases are
made:

 1) Reinstante the waiver right away
 2) Reinstante the waiver when a compatibility with the post Avocado
    98.0 release (the development version of Avocado) is found.

I'd suggest going with 2, and creating issues to track the
compatibility problems found and that need to be addressed before the
next compatible release.

Signed-off-by: Cleber Rosa <crosa@redhat.com>