CI: increase integration with test providers coverage
By expanding the number of tests from test providers that we run, we
increase the level of confidence that Avocado-VT API changes are not
breaking tests.

The list of tests added was curated by attempting to add tests that
are known to be stable on other testing scenarios, but adapted to the
Cirrus CI environment (based on JeOS, different KVM host environment,
etc).

This list is not set in stone, and can and should be continually
curated to represent a better sampling of tests and thus a better
integration coverage.

Signed-off-by: Cleber Rosa <crosa@redhat.com>