CI: include the path of avocado-runner-avocado-vt on list
This is a workaround for the fact that current Avocado will look for
the runner script while doing resolution (creating Runnables) because
it will populate the Runnable configuration with the "used
configuration" for that kind of runner.

If a change in Avocado will be needed is TBD, but this takes care of
the broken resolution in the CI environment.

Reference: https://github.com/avocado-framework/avocado-vt/issues/3457
Signed-off-by: Cleber Rosa <crosa@redhat.com>