Update "--nvram" to "--keep-nvram" during undefine
For UEFI guest, we can not undefine it successfully without any nvram
options. With "--nvram", it will delete the nvram file when undefine the
guest. When we define the guest again, it will reset before normal
boot. As serail login will input something during the reset process, it
will trigger the guest wait for some manual intervention which cause auto
case to fail. Update "--nvram" to "--keep-nvram" to workaround it.

Signed-off-by: Yalan Zhang <yalzhang@redhat.com>