virttest/migration: improve error message when do_cancel fails
If migration is not running (crashed or completed) when trying
to cancel, we'll get a an error from further down in the code,
more specifically from the type cast in `process.safe_kill`.

I've observed two possible test errors:
1. ValueError: invalid literal for int() with base 10: ''
2. TypeError: int() argument must be a string, a bytes-like object or a number, not 'NoneType'

However, when we call cancel we usually want to test the behavior
of cancellation and possibly retrying. In these cases, now print
a specific error message about the migration not running.

If the test log shows the migration had already finished, we
can try to slow down the migration with things like decreasing
the bandwidth or creating dirty memory pages.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>