Fix the EMPTY_STRING issue
When we set 'EMPTY_STRING' to an option(e.g. romfile=EMPTY_STRING)
VT assigns '""' to the option, e.g. romfile = '""', this can work
fine in the qemu cmdline in raw format
e.g. -device virtio-net-pci,romfile="",
but that never works for the cmdline in json format, for "" is
recognized as a regular string instead of a null string, e.g.
{'execute': 'device_add', 'arguments': {'romfile': '""', ...}},
which leads to 'failed to find romfile """"' error.

Signed-off-by: Zhenchao Liu <zhencliu@redhat.com>