Add read permission to a bit of log files
Let me take libvirtd.log as the example. There is an issue reported to
libvirt project: https://gitlab.com/libvirt/libvirt/-/issues/71, the use
case is people will use non-root user to analyze the libvirt instance
log file in another system, but by default the permission of libvirtd
log files are 0600, which reject other users to open the log for
analyzing.

For logs collection in testing, they should not contain confidential
information or not exposed to the outside, so add read permission for
all users should not a dangerous operation.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>