Update vm_xml.sync() method to correctly use nvram
Previously sync method was calling undefine method based with the xml
that was given in argument. However when we were undefining a VM and the
vmxml did not have nvram in os part of xml, the option "--nvram" was not
added and the undefine failed.

This commit changes that and nvram is looked up in the xml that is
dumped from actively used xml. So if we're undefining a VM with nvram
the --nvram option will be added and a VM will be correctly undefined.

Signed-off-by: Vaclav Hodina <vhodina@redhat.com>