Allow destroying vms via monitor upon recreation
In some cases the vm's memory and its images are known to be clean
of corruption and we can save up to 2 minutes per test (for thousands
of tests) if we destroy the vm faster before recreating it anew.

We can use this when restoring the vm from a file as we make sure
the vm is paused before recreating it and thus no vm image will be
harmed in the process.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>