Configure vm.destroy gracefulness for performance when discarding vms
During environment setup we discard vms that are not going to
participate in a test but doing so gracefully results in a large
timeout (60 seconds by default) for each one and slows down the
test and overall test run significantly. As for some plugins the
discarded vm will be fully restored from a clean state when
it is needed again by a test, graceful shutdown is not always
needed and ridding ourselves of it will improve performance for
free in these cases. All default cases can still not do this to
keep their reused images clean.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>