Wait for proper monitor status after any memory file restoration
It seems that if the vm is still in "paused (inmigrate)" status and
similar, the previous check will consider this successfully paused
and the restoration thus fully completed. Any later resuming of the
vm then will have its monitor completely ignoring the "cont" monitor
command and thus the vm will remain paused indefinitely.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>