Allow the option of creating a vm firstly after loading from file
Loading a running vm's memory from file would require it to be paused
first but would also work just fine without any prior vm. In some cases
the user has to create a vm and pause it just to be able to satisfy the
requirements of previously running and paused vm thus leading to overhead
and circumventing more direct use of this functionality.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>