qemu_vm: Fix swtpm path too long for vTPM + ping-pong migration
The current workaround is adding "_copied" string into the swtpm path
for vTPM migration. When we do ping-pong migration few times, the socket
path is longer than system limitation(107), which raises "Path for
UnioIO socket is too long". But actually the even times' migration can
use the source vTPM socket, so here add an action to roll back the tpm.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>