qdevices: introduce `QMachine` and apply it to virttest
It has been proved that using `QStrDev` to represent '-machine' was
not a good choice nowadays, more and more workarounds have to be
applied to manipulate the machine device properly.

This patch introduces `QMachine`, the dedicated representation for
machine devices, to resolve this issue. Also, updated the related
code accordingly.

Signed-off-by: Xu Han <xuhan@redhat.com>