Fix incremental backup data comparison failures
Previously the data will be dumped to /path/to/original_file.dmp.
But when the original disk is a block device, such as /dev/sdb,
the data will be dumped to /dev/sdb.dump which is not a suitable place.
Or if the origial image is in a readonly dir, of course the data
could not be dumped to that dir. This patch is to fix above issues
to use the tmp dir to save the dumped data when do data comparison.

Signed-off-by: Yi Sun <yisun@redhat.com>