Add conditional support for loader based plugins
The "loader" architecture has been removed in the latest Avocado
commits (and will be reflected on Avocado version 100.0 and later).

This adds a conditional support for the loader based plugins, so that
the latest Avocado-VT can still be used (hopefully for a short time)
with Avocado versions having loader support.

Fixes: https://github.com/avocado-framework/avocado-vt/issues/3591
Signed-off-by: Cleber Rosa <crosa@redhat.com>