Raise VMCreateError if QEMU VM is dead when connecting to monitor
`create` method from `qemu_vm.py` tries to connect to QMP monitor right after
initial cheking of QEMU process to be alive. This can bring timing issue as
sometimes QEMU process can die right after this initial check when connections
to QMP monitors are being established.

This commit adds check for status of process when `MonitorConnectError`
exception is raised. If the QEMU VM is dead, then `VMCreateError` exception
is raised instead, so the true reason of failure is not hidden.

Signed-off-by: Lukas Kotek <lkotek@redhat.com>