Fix acceptable negative value range for cfs_quota params
For both cgroup v1 and v2, all negative int are acceptable
for the cfs_quota params, such as vcpu_quota, emulator_quota,
global_quota and iothread_quota. And a negative value means the
param is set as 'max' (or 'unlimited'). So we should not only
match '-1' to 'max', but match negative int to 'max'.

Signed-off-by: Yi Sun <yisun@redhat.com>