HugePageConfig:Extended multiple huge page settings
Extended multiple huge page settings methods,
Allow users to customize the selection of huge page sizes for testing.
this patch will keep the default settings,
and add new methods to allow users to control the page size.
The default value method is used to ensure compatibility
when the user does not configure.

Signed-off-by: zhenyzha <zhenyzha@redhat.com>