verify_kernel_crash:Expand the scope of regularity checking
According to the comment left by Linus Torvalds in the development
community "We're trying to get rid of BUG_ON() because it kills the
machine. Not replace it with another bogus thing that kills a machine."

Therefore, in future kernel development, developers will not
use BUG_ON() unless necessary.
On the other hand, the expansion of the detection range is a
double-edged sword.
I don't want to classify all end traces as kernel crashes.
So only widen the search if console has a complete "cut here"
"end trace", remove the "BUG:" keyword.

Signed-off-by: zhenyzha <zhenyzha@redhat.com>