test_setup: Refactor network proxies setup
This is the first commit of a series that will try to tidy up the
env_process module's preprocess and postprocess methods.

The first refactored element is the network proxies setup. This commit
adds the NetworkProxies setuper. The latter will take care of the needed
actions to prepare the test env according to the configuration params.
It also registers it in the setup_manager already defined in the
env_process module.

Signed-off-by: bgartzi <bgartzia@redhat.com>