KSMConfig.cleanup: Ensure ksm to be disabled at the end of the case
In the `cleanup` function, "start_ksmtuned" will start the
ksmtuned.service, it controls and tunes the ksm service, but depends on
the KSM_MONITOR_INTERVAL. So ksm is not closed immediately, but at some
time in KSM_MONITOR_INTERVAL loop. Without this fix, the next test case
may launch the guest with ksm enabled, which is not expected.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>