utils_disk/libvirt_disk: add way to handle disk by path
Partitions in /proc/partitions are not persistent which
can lead to problems when a test checks if a disk was cold
plugged.

Use persistent naming from /dev/disk/by-path which allows
us to easily identify those disks and partitions with
the attached disk by their libvirt <address/>.

Add and update other functions to be compatible with this way
of dealing with partitions.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>