qemu_vm: Split the life cycle management of VM into finer granularity
The current "create" method actually consists of qemu command line
generation and qemu process creation. There are also "destroy" methods
that include stopping the qemu process and cleaning up resources.

However, in th VM life cycle, we should be able to custom the qemu
command line before launch it, right now we use `extra_params` but
it's not a good idea, with this patch, we can deprecat it, but insert
some devices/config which managed by qcontainer.DevContainer.

Besides this, after the VM is stopped, we should be able to use the
existing resources to restart the qemu process, without VM definition.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>