Unify all logfile handling into utils_logfile
Most importantly, for better decoupling with aexpect we provide
sessions with close hooks that should have the same effect of
closing the open log file descriptors but in a more decoupled
fashion.

The logging functionality in utils_misc has been fully deprecated
for the sake of the same functionality in a fully devoted module.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>