fix: Empty "image_size" in sn_params when "image_snapshot = yes"
In qcontainer.py, when creating snapshot images, will get the image size
from params. But if the image_size is defined in different variants, the
sn image size may smaller than the base image, which is incorrect. The
reasonable way is empty the image_size in sn_params, to make the size of
snapshot image align to the base one, as this is just a temp snapshot.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>