fix(utils_logfile.log_line): Open file with 'a' mode for safer writing
utils_logfile.log_line currently uses 'w' mode to open log files, this
causes some log contents missed when closing the log file. For example,
when we close the log file, there are some contents in the cache and
waiting for writing to the log, so need to reopen it, if use 'w' mode,
old contents will be truncated.

Suggested-by: Lukáš Doktor <ldoktor@redhat.com>
Signed-off-by: Yihuang Yu <yihyu@redhat.com>