vm_xml: set cpu mode
On x86_64, if we set <cpu>xxx</cpu> in vm xml and save, then libvirt could automatically
fill in the cpu mode. But on arm now, libvirt can not make it which cause vm fails to
boot with below error. Libvirt dev is working on cpu model, so it is not good time to
support this auto filling now. So we need a workaround here to set it explicitly.

Signed-off-by: Dan Zheng <dzheng@redhat.com>