USB: fix Invalid value for attribute 'bus' in element 'address'
From libvirt document, The values of these attributes can be given
in decimal, hexadecimal (starting with 0x) or octal (starting with 0)
form.

In our script, the analyze of 'lsusb' output pass bus and address
literally. If the bus num is greater or equal than 8, for instance
008', libvirt will report the error.

In the lsusb output, the bus value should be decimal, so just remove
the leading zeros from the string to fix it.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>