env_process: Fix the issue of iterating an integer
A recent avocado commit(241c9cdf62f0162e24c37913ab37053971563ee4)
enhanced cpu.get_family to return an integer for a AMD cpu, which
can cause error, but get_family was declared to return a *name*,
instead of an integer, which is confused. Added this commit as a
workaround to bypass the error.

Signed-off-by: Zhenchao Liu <zhencliu@redhat.com>