cpu: adapt help text parsing to support modern qemus
Starting from qemu v9.0.0, the help text of cpu models associated
with all the hardware architectures was unified to the pattern
below.

    $ qemu -cpu '?'
    Available CPUs:
      <model>   <description>
      ...

As a result, we have to modify the corresponding parseing function
to cater for the change, since the arch names were no longer showed
in front of the model names.

This patch is going to introduce a new parser for handling the new
pattern of cpu help text, meanwhile keep the old one for supporting
legacy qemus.

Signed-off-by: Xu Han <xuhan@redhat.com>