env_process: relocate migrate vm registration
In the preprocess routine from env_process a piece of code is in charge
of registering the "migrate_vms" into the "vm" configuration parameter.
Some of the environment preprocessing steps might need from that
configuration to work properly. However, due to the ongoing refactoring
of the env_process:{pre,post}process functions, the migrate_vms might be
registered after the environment setup step. This will probably make
some of the setup steps misbehave.

For that reason, let's move the migrate_vm registration step to the
beginning of the preprocess function.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>