Enable virt_use_nfs when selinux is enforcing
In function setup_or_cleanup_nfs, when selinux is enforcing, it only
prints a debug message asking libvirt to enable virt_use_nfs, but
doesn't exit. This hides a bug when in CI env, the virt_use_nfs may be
set by other scripts or not, so another script using the function may
fail randomly if it doesn't set virt_use_nfs as required.

Since virt_use_nfs is a very common operation, it's fine to enable
it here.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>