utils-test:Add running htx-cmdline stress_type
This patch can be used to run htx mdts stress_type
using htxcmdline tool
Additionally, this patch fixes DeprecationWarning of
invalid escape sequence 'x' W605

This patch introduces the following changes:
1. Allow running of different stress types in different
vms based on stress_type_%s parameter if provided in
the cfg file, where %s is the name of the vm. This will
allow us to introduce additional types of stress in the
future. If the stress_type_%s paramter is not provided,
then the default stress_type parameter will be used.

2. define a new function load_htxstress_tool() to run
htxcmdline tool in the guest. This function will be
called whenever stress_type_vmname = htxcmdline. The user
can provide the link to the .rpm/.deb package and the
function will handle the installation, followed by the
running of the htxcmdline tool by using the htxcmdline_args
parameter, all provided by user in the cfg file.

Signed-off-by: Misbah Anjum N <misanjum@linux.vnet.ibm.com>