env_process: Refactor host dmesg verification
Host dmesg can be checked if the test configuration requires to do so.
It will be in fact be checked both before and after running the test.

However, that was done directly in the preprocess and postprocess
functions in virttest.env_process. Write a Setuper subclass that
implements those in setup/cleanup methods and register the setuper in
the env_process setup_manager.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>