env_process: Refactor kvm module reload/restore steps
KVM modules are reloaded and restored before and after running tests.
That was done directly in the preprocess and postprocess functions in
virttest.env_process. Write a Setuper subclass that implements those in
setup/cleanup methods and register the setuper in the env_process
setup_manager.

This is a patch from a larger patch series refactoring the env_process
preprocess and postprocess functions. In each of these patches, a
pre/post process step is identified and replaced with a Setuper subclass
so the following can finally be met:
    - Only cleanup steps of successful setup steps are run to avoid
      possible environment corruption or hard to read errors.
    - Running setup/cleanup steps symmetrically during env pre/post
      process.
    - Reduce explicit pre/post process function code length.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>