env_process: make version_info a global submodule variable
This is part of a larger patch series that will refactor env_process
preprocess steps that verify tested package versions.

In this patch, version_info is made a global variable of env_process
so it can be imported from other modules. This will help carry out
the refactoring in small steps rather than in a big unique one.

This will change on the last patch of this series. This is only to make
sure that each commit in this series work properly.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>