env_process: move version_info into global requirement_checks variable
This is part of a larger patch series that will refactor env_process
preprocess steps that verify tested package versions.

In this patch, version_info is made a global variable of
test_setup.requirement_checks so it is available in the scope in which
package version checks are going to be rewritten.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>